requirements.txt
requests
pyperclip
playwright
PyYAML
soundcard
soundfile
future
numpy
librosa
tensorflow
pillow
docopt
joblib
scipy
psutil
flask